requirements.txt
numpy==1.26.4
onnxruntime==1.19.2
pillow==10.3.0
rembg==2.0.61
gradio
opencv-python-headless==4.8.1.78
torch==2.2.0
torchvision==0.17.0 